# server/openapi/openapi.yaml
...
openapi: 3.1.0
info:
  title: Loventia API
  version: 1.0.0
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  contact:
    name: Loventia API Team
    url: https://loventia.example.com
    email: api@loventia.example.com
  description: |
    OpenAPI specification for the Loventia dating application.

    This spec is aligned with the current server/src/app.js routing we just test-drove
    with PowerShell:

      - Primary auth endpoint: POST /api/auth/login
      - Legacy/fallback auth endpoint: POST /api/users/login
      - Refresh: POST /api/auth/refresh
      - Self: GET /api/auth/me and GET /api/users/me and GET /api/me
      - Reset-password flow: POST /api/auth/forgot-password, POST /api/auth/reset-password
      - Private diagnostics: GET /api/auth/private/ping, GET /api/auth/private/no-me
      - User photo management (the things we just tested): 
          * DELETE /api/users/{userId}/photos/{slot}
          * DELETE /api/users/{userId}/photos?path=...
          * PUT    /api/users/{userId}/photos/reorder
          * POST   /api/users/{userId}/set-avatar
      - Billing endpoints (Stripe webhook at /webhooks/stripe and legacy /api/payment/stripe-webhook)
      - Admin CSV exports
      - Health/readiness endpoints

    IMPORTANT:
      - We **document both** logins on purpose:
          1) /api/auth/login  (main, preferred)
          2) /api/users/login (legacy/fallback)
      - This matches the README change you will add in server/README.md
servers:
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    ProfileCard:
      type: object
      description: Public information displayed in Discover listings.
      properties:
        id:
          type: string
          example: 68dab8a761a412065159ff4d
        username:
          type: string
          example: Ville1981
        age:
          type: integer
          example: 43
        gender:
          type: string
          example: male
        bio:
          type: string
          example: Coffee, synthwave and long runs by the sea.
        distanceKm:
          type: number
          example: 12.5
        profilePicture:
          type: string
          example: /uploads/1759343842730-848316731.gif
        photos:
          type: array
          items:
            type: string
          example:
            - /uploads/p1.jpg
            - /uploads/p2.jpg
        location:
          type: object
          properties:
            lat:
              type: number
              example: 60.1699
            lng:
              type: number
              example: 24.9384
    Pagination:
      type: object
      properties:
        page:
          type: integer
          minimum: 1
          example: 1
        limit:
          type: integer
          minimum: 1
          example: 20
        total:
          type: integer
          minimum: 0
          example: 237
        hasNext:
          type: boolean
          example: true
        hasPrev:
          type: boolean
          example: false
...
paths:
  /api/discover:
    get:
      tags: [Discover]
      summary: Discover profiles
      description: Returns a paginated list of public profile cards.
      operationId: discover_list
      security: [{ bearerAuth: [] }]
      parameters:
        - in: query
          name: q
          schema: { type: string }
        - in: query
          name: minAge
          schema: { type: integer, minimum: 18 }
        - in: query
          name: maxAge
          schema: { type: integer, minimum: 18 }
        - in: query
          name: gender
          schema: { type: string }
        - in: query
          name: distanceKm
          schema: { type: number }
        - in: query
          name: lat
          schema: { type: number }
        - in: query
          name: lng
          schema: { type: number }
        - in: query
          name: page
          schema: { type: integer, minimum: 1, default: 1 }
        - in: query
          name: limit
          schema: { type: integer, minimum: 1, default: 20 }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  users:
                    type: array
                    items:
                      $ref: '#/components/schemas/ProfileCard'
                  meta:
                    $ref: '#/components/schemas/Pagination'
        '400': { description: Invalid query parameters }
        '401': { description: Unauthorized }
...
  /__routes:
    get:
      tags: [Diagnostics]
      summary: List app-level routes (dev-only)
      description: Returns app-level mounted routes (dev-only).
      operationId: diagnostics_routes
      responses:
        '200':
          description: List of routes
        '401':
          description: Unauthorized
...
security:
  - bearerAuth: []